Add 'repeat' option to replace()
This repeatedly applies replace until the final expression no longer
changes. This is useful for replacement inside TensExpr (since one may
have multiple matches for the query in the same TensExpr, and each
invocation of replace will replace only one of those).

For repeat=False (default), the behaviour of the function is unchanged)